+++ tsup.config.ts
@@ -1,11 +1,12 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-  name: 'tsup',
   target: 'es2019',
   dts: true,
   clean: true,
   sourcemap: true,
   format: ['esm', 'cjs'],
-  treeshake: 'smallest',
+  banner: {
+    js: "'use client'",
+  },
 })